Adding module for Security group
Resource security group was removed from main file of this project and included in new module for security group.

Currently, it is not so flexible: you can only change vpc_id and name. But in future I plan to improve flexibility by creating in this module egress and ingress rules.